Revert "int to size_t"
This reverts commit 94c9e9fb99f16285aa8f200b864c8746565d3132.

+++ Utilities/Communication/comm_resize.c
@@ -37,11 +37,11 @@ int      *sum_recv_sizes)       /* sum of the sizes of the items I'll receive */
     int      *sort_val = NULL;	/* start_from to order properly */
     int      *sizes_to = NULL;  /* size of each msg to send (if items vary) */
     int      *sizes_from = NULL;/* size of each msg to recv (if items vary) */
-    size_t   *starts_to_ptr = NULL;	/* where in dense vector sends start */
-    size_t   *starts_from_ptr = NULL;	/* where in dense vector recvs start */
-    size_t   *indices_to_ptr = NULL;	/* where to find items I send */
+    int      *starts_to_ptr = NULL;	/* where in dense vector sends start */
+    int      *starts_from_ptr = NULL;	/* where in dense vector recvs start */
+    int      *indices_to_ptr = NULL;	/* where to find items I send */
 					/* ordered like procs_to */
-    size_t   *indices_from_ptr = NULL;	/* where to find items I recv */
+    int      *indices_from_ptr = NULL;	/* where to find items I recv */
 					/* ordered like procs_from */
     int       var_sizes;        /* items have variable sizes? */
     int       i, j, k;		/* loop counters */
@@ -122,7 +122,7 @@ int      *sum_recv_sizes)       /* sum of the sizes of the items I'll receive */
 		need to allocate, set indices_to_ptr
 	   3,4. mirror cases for _from
 	*/
-	starts_to_ptr = (size_t *) ZOLTAN_MALLOC((nsends + self_msg) * sizeof(size_t));
+	starts_to_ptr = (int *) ZOLTAN_MALLOC((nsends + self_msg) * sizeof(int));
 	if (starts_to_ptr == NULL && (nsends + self_msg) != 0) {
 	    return_flag = ZOLTAN_MEMERR;
 	    goto Mem_Err;
@@ -167,7 +167,7 @@ int      *sum_recv_sizes)       /* sum of the sizes of the items I'll receive */
 
 	else {		/* Harder case, sends not blocked */
 	    offset = (int *) ZOLTAN_MALLOC(plan->nvals * sizeof(int));
-	    indices_to_ptr = (size_t *) ZOLTAN_MALLOC(plan->nvals * sizeof(size_t));
+	    indices_to_ptr = (int *) ZOLTAN_MALLOC(plan->nvals * sizeof(int));
 	    if ((offset == NULL || indices_to_ptr == NULL) && plan->nvals != 0) {
 	        return_flag = ZOLTAN_MEMERR;
 	        goto Mem_Err;
@@ -206,7 +206,7 @@ int      *sum_recv_sizes)       /* sum of the sizes of the items I'll receive */
 	    sizes_from, plan->procs_from, nrecvs, 
 	    &plan->total_recv_size, my_proc, tag, plan->comm);
 
-	starts_from_ptr = (size_t *) ZOLTAN_MALLOC((nrecvs + self_msg) * sizeof(size_t));
+	starts_from_ptr = (int *) ZOLTAN_MALLOC((nrecvs + self_msg) * sizeof(int));
 	if (starts_from_ptr == NULL && (nrecvs + self_msg) != 0) {
 	    return_flag = ZOLTAN_MEMERR;
 	    goto Mem_Err;
